backend/pyproject.toml
[tool.poetry]
name = "esprit-backend"
version = "0.1.0"
description = "Esprit Backend Orchestrator Service"
authors = ["Esprit"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
httpx = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
boto3 = "^1.35.0"
supabase = "^2.10.0"
litellm = "^1.79.0"
redis = "^5.2.0"
structlog = "^24.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
